Changes datatype & history module
- ICNum int change to bigInt in phpmyadmin db
- ICNum int change to long in servlet
- History only preview 'Approved' or 'Declined' program
- Update cmmsdb included

+++ CMM_System/src/java/programStudent/stud_merit_GETapprovalProgram.java
@@ -98,7 +98,7 @@ public class stud_merit_GETapprovalProgram extends HttpServlet {
 //                dispmerit.setProgID(progid);
                 dispmerit.setRole(rs1.getString("role"));
                 dispmerit.setMatricNum(rs1.getString("matricNum"));
-                dispmerit.setIcNum(rs1.getInt("icNum"));
+                dispmerit.setIcNum(rs1.getLong("icNum"));
                 dispmerit.setMerit(rs1.getInt("merit"));
                 
                 dm.addElement(dispmerit);
